+++ src/episode_04/InputDemo.java
@@ -27,7 +27,7 @@
  * either expressed or implied, of the FreeBSD Project.
  */
 
-package episode_4;
+package episode_04;
 
 import org.lwjgl.LWJGLException;
 import org.lwjgl.input.Keyboard;
