+++ src/episode_07/TimerDemo.java
@@ -1,0 +1,187 @@
+/*
+ * Copyright (c) 2012, Oskar Veerhoek
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ * The views and conclusions contained in the software and documentation are those
+ * of the authors and should not be interpreted as representing official policies,
+ * either expressed or implied, of the FreeBSD Project.
+ */
+
+package episode_07;
+
+import org.lwjgl.LWJGLException;
+import org.lwjgl.Sys;
+import org.lwjgl.opengl.Display;
+import org.lwjgl.opengl.DisplayMode;
+
+import static org.lwjgl.opengl.GL11.*;
+
+/**
+ * Shows framerate-independent movement
+ *
+ * @author Oskar
+ */
+public class TimerDemo {
+
+    private static long lastFrame;
+
+    private static long getTime() {
+        return (Sys.getTime() * 1000) / Sys.getTimerResolution();
+    }
+
+    private static double getDelta() {
+        long currentTime = getTime();
+        double delta = (double) (currentTime - lastFrame);
+        lastFrame = getTime();
+        return delta;
+    }
+
+    public static void main(String[] args) {
+        try {
+            Display.setDisplayMode(new DisplayMode(640, 480));
+            Display.setTitle("Timer Demo");
+            Display.create();
+        } catch (LWJGLException e) {
+            e.printStackTrace();
+            Display.destroy();
+            System.exit(1);
+        }
+
+        int x = 100;
+        int y = 100;
+        int dx = 1;
+        int dy = 1;
+
+        // Initialization code OpenGL
+        glMatrixMode(GL_PROJECTION);
+        glLoadIdentity();
+        glOrtho(0, 640, 480, 0, 1, -1);
+        glMatrixMode(GL_MODELVIEW);
+
+        lastFrame = getTime();
+
+        while (!Display.isCloseRequested()) {
+            // Render
+
+            glClear(GL_COLOR_BUFFER_BIT);
+
+            double delta = getDelta();
+            x += delta * dx * 0.1;
+            y += delta * dy * 0.1;
+
+            glRecti(x, y, x + 30, y + 30);
+
+            Display.update();
+            Display.sync(60);
+        }
+
+        Display.destroy();
+        System.exit(0);
+    }
+}
+
+interface Entity2D {
+    public float getX();
+
+    public float getY();
+
+    public void setX(float x);
+
+    public void setY(float y);
+
+    public void setLocation(float x, float y);
+
+    public void setUp();
+
+    public void destroy();
+
+    public void draw();
+}
+
+abstract class AbstractEntity2D implements Entity2D {
+    protected float x;
+    protected float y;
+
+    public float getX() {
+        return x;
+    }
+
+    public float getY() {
+        return y;
+    }
+
+    public void setX(float x) {
+        this.x = x;
+    }
+
+    public void setY(float y) {
+        this.y = y;
+    }
+
+    public void setLocation(float x, float y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public abstract void setUp();
+
+    public abstract void destroy();
+
+    public abstract void draw();
+}
+
+class Box2D extends AbstractEntity2D {
+    protected float size;
+
+    public Box2D(float x, float y, float size) {
+        this.x = x;
+        this.y = y;
+        this.size = size;
+    }
+
+    public Box2D() {
+        this.x = this.y = this.size = 0;
+    }
+
+    @Override
+    public void setUp() {
+        // We don't need anything here for a box
+    }
+
+    @Override
+    public void destroy() {
+        // We don't need anything here for a box
+    }
+
+    @Override
+    public void draw() {
+        glBegin(GL_TRIANGLES);
+        glVertex2f(x + size / 2, y + size / 2);
+        glVertex2f(x + size / 2, y - size / 2);
+        glVertex2f(x - size / 2, y - size / 2);
+        glVertex2f(x - size / 2, y - size / 2);
+        glVertex2f(x - size / 2, y + size / 2);
+        glVertex2f(x + size / 2, y + size / 2);
+        glEnd();
+    }
+}
